TOA-13. Package structure changed, added DataResponse.kt for network requests
