+++ app/src/main/java/ru/fursa/toa/core/data/DataResponse.kt
@@ -1,0 +1,6 @@
+package ru.fursa.toa.core.data
+
+sealed class DataResponse<out T> {
+    data class Success<out T>(val data: T): DataResponse<T>()
+    data class Error<out T>(val error: Throwable): DataResponse<T>()
+}
