+++ app/src/main/java/ru/fursa/toa/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package ru.fursa.toa.ui.theme
+package ru.fursa.toa.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
