+++ app/src/main/java/ru/fursa/toa/login/domain/model/LoginResponse.kt
@@ -1,0 +1,5 @@
+package ru.fursa.toa.login.domain.model
+
+data class LoginResponse(
+    val authToken: String,
+)
